feat(Review): ReviewScore 데이터 타입 변경 및 @Valid 추가
- 5점 척도 평가를 위한 reviewScore의 데이터 타입 변경(Long -> Integer)
- Review 생성 시, 적합한 값을 받아오기 위하여 @Valid 추가 및 적용

+++ review_service/src/main/java/org/personal/review_service/domain/Review.java
@@ -22,7 +22,7 @@ public class Review {
     private String reviewContent;
 
     @Column(name = "review_score", nullable = false)
-    private Long reviewScore;
+    private Integer reviewScore;
 
     @Column(name = "review_registred_date", nullable = false)
     private String reviewRegisteredDate;
@@ -37,7 +37,7 @@ public class Review {
     private Long locationId;
 
     @Builder
-    public Review(Long reviewId, String reviewContent, Long reviewScore, String reviewRegisteredDate,
+    public Review(Long reviewId, String reviewContent, Integer reviewScore, String reviewRegisteredDate,
                   Boolean reviewIsDeleted, Long userId, Long locationId) {
         this.reviewId = reviewId;
         this.reviewContent = reviewContent;
